+++ core/src/main/scala/scalaz/Semigroup.scala
@@ -118,5 +118,12 @@ object Semigroup {
   def iterate[F[_], A](a: A)(f: A => A)(implicit F: Applicative[F], m: Semigroup[F[A]]): F[A] =
     m.append(F.point(a), iterate[F, A](f(a))(f))
 
+  /** Semigroup is an invariant functor. */
+  implicit val semigroupInvariantFunctor: InvariantFunctor[Semigroup] = new InvariantFunctor[Semigroup] {
+    def xmap[A, B](ma: Semigroup[A], f: A => B, g: B => A): Semigroup[B] = new Semigroup[B] {
+      def append(x: B, y: => B): B = f(ma.append(g(x), g(y)))
+    }
+  }
+
   ////
 }
